[SBL-46] 커버리지 100
