+++ src/test/kotlin/com/sbl/sulmun2yong/drawing/domain/DrawingHistoryTest.kt
@@ -1,36 +1,18 @@
 package com.sbl.sulmun2yong.drawing.domain
 
-import com.sbl.sulmun2yong.global.data.PhoneNumber
+import com.sbl.sulmun2yong.fixture.drawing.DrawingHistoryFixtureFactory
+import com.sbl.sulmun2yong.fixture.drawing.DrawingHistoryFixtureFactory.drawingBoardId
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Test
-import java.util.UUID
 
 class DrawingHistoryTest {
     @Test
     fun `추첨 기록을 만든다`() {
         // given
-        val participantId = UUID.randomUUID()
-        val phoneNumber = PhoneNumber("010-1234-5678")
-        val drawingBoardId = UUID.randomUUID()
-        val selectedTicketIndex = 0
-        val ticket =
-            com.sbl.sulmun2yong.drawing.domain.ticket.Ticket.Winning(
-                "테스트 아메리카노",
-                "테스트",
-                false,
-            )
-        // when
-        val drawingHistory =
-            DrawingHistory.create(
-                participantId,
-                phoneNumber,
-                drawingBoardId,
-                selectedTicketIndex,
-                ticket,
-            )
+        val drawingHistory = DrawingHistoryFixtureFactory.createdDrawingHistory()
 
-        // then
+        // when, then
         with(drawingHistory) {
             assertNotNull(id)
             assertEquals(participantId, this.participantId)
@@ -38,6 +20,29 @@ class DrawingHistoryTest {
             assertEquals(drawingBoardId, this.surveyId)
             assertEquals(selectedTicketIndex, this.selectedTicketIndex)
             assertEquals(ticket, this.ticket)
+        }
+    }
+
+    @Test
+    fun `추첨 기록을 그룹을 만든다`() {
+        // given
+        val drawingHistory1 = DrawingHistoryFixtureFactory.createdDrawingHistory()
+        val drawingHistory2 = DrawingHistoryFixtureFactory.createdDrawingHistory()
+        val drawingHistory3 = DrawingHistoryFixtureFactory.createdDrawingHistory()
+
+        // when
+        val drawingHistoryGroup =
+            DrawingHistoryGroup(
+                surveyId = drawingBoardId,
+                count = 3,
+                histories = listOf(drawingHistory1, drawingHistory2, drawingHistory3),
+            )
+
+        // then
+        with(drawingHistoryGroup) {
+            assertEquals(drawingBoardId, this.surveyId)
+            assertEquals(3, this.count)
+            assertEquals(3, this.histories.size)
         }
     }
 }
